pyproject.toml
[tool.poetry]
name = "picasso-test-task"
version = "0.1.0"
description = ""
authors = ["Irina Fok <fokira@mail.ru>"]
readme = "README.md"
packages = [{include = "picasso_test_task"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
redis = "^5.0.0"
celery = "^5.3.1"
python-dotenv = "^1.0.0"
pytest = "^7.4.2"
pytest-django = "^4.5.2"
factory-boy = "^3.3.0"
pytest-mock = "^3.11.1"

[tool.poetry.group.dev.dependencies]
drf-yasg = "^1.21.6"
flake8 = "^6.0.0"
isort = "^5.12.0"
django-debug-toolbar = "^4.1.0"
pre-commit = "^3.3.3"
black = "^23.3.0"
pep8-naming = "^0.13.3"
flake8-broken-line = "^1.0.0"
flake8-return = "^1.2.0"
flake8-isort = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
